+++ engine/graphics/src/vulkan/graphics_vulkan_device.cpp
@@ -297,7 +297,7 @@ namespace dmGraphics
 
         HashState64 shader_hash_state;
         dmHashInit64(&shader_hash_state, false);
-        dmHashUpdateBuffer64(&shader_hash_state, source, sourceSize);
+        dmHashUpdateBuffer64(&shader_hash_state, source, (uint32_t) sourceSize);
         shaderModuleOut->m_Hash = dmHashFinal64(&shader_hash_state);
 
         return vkCreateShaderModule( vk_device, &vk_create_info_shader, 0, &shaderModuleOut->m_Module);;
